feat(playground): host create-knowledge flow inside MCPOverlay
Extract the create-knowledge form into NewKnowledgeFormBody so it can render
either inside a standalone Dialog or as a routed view of another modal. The
standalone NewKnowledgeOverlay still wraps the body for callers that aren't
already in a modal (e.g. WorkspaceForm).

MCPSelector gains an optional onRequestCreateKnowledge prop. When provided
(by MCPOverlay), the "+" button calls it instead of opening the standalone
Dialog, so we don't stack Dialogs. When omitted (WorkspaceForm), the local
Dialog flow still works.

MCPOverlay adds a view: "list" | "create" state. In create view it swaps to
create-specific header/footer chrome and hosts NewKnowledgeFormBody. On
success the new knowledge MCP is folded into the draft and view returns to
the list with the KNOWLEDGE tab active. Dismissal is blocked while a create
is in flight so the draft can't be lost.

Caller fixes for the new flow's embedding-engine EngineSelect: pass
justify-start so the trigger text left-aligns, and popoverContentProps with
align="start" and trigger-width binding so the dropdown matches the page.
Adds px-1 py-1 to the scroll wrapper so focus rings on inputs aren't clipped
by overflow-y-auto.
