+++ packages/playground/src/components/knowledge/index.ts
@@ -1,2 +1,3 @@
 export * from "./embed-documents-overlay";
+export * from "./new-knowledge-form-body";
 export * from "./new-knowledge-mcp-overlay";
